# ring: add GetWithOptions method to adjust per call behavior