@56quarters 2024-12-18
I wasn't able to avoid these functions escaping to the heap. Benchmarking the two methods shows that `GetWithOptions` is about 14% slower. I'm going to leave these as mostly separate methods and work on adding tests to `GetWithOptions`. After that I'll mark this as ready for review and we can start experimenting with higher replication factors in Mimir.

```
$ benchstat old.txt new.txt 
goos: linux
goarch: amd64
pkg: github.com/grafana/dskit/ring
cpu: Intel(R) Core(TM) i9-10885H CPU @ 2.40GHz
                                                         │   old.txt    │                new.txt                 │
                                                         │    sec/op    │    sec/op      vs base                 │
Ring_Get/without_zone_awareness-16                         863.2n ± ∞ ¹   1009.0n ± ∞ ¹        ~ (p=1.000 n=1) ²
Ring_Get/without_zone_awareness,_not_enough_instances-16   654.2n ± ∞ ¹    758.2n ± ∞ ¹        ~ (p=1.000 n=1) ²
Ring_Get/with_zone_awareness-16                            1.138µ ± ∞ ¹    1.252µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Ring_Get/one_excluded_zone-16                              855.0n ± ∞ ¹    962.0n ± ∞ ¹        ~ (p=1.000 n=1) ²
geomean                                                    861.0n          979.7n        +13.80%
```