@56quarters 2024-12-30
After adding unit tests for this change, it turns out to not be compatible with zone-awareness. There's an assumption in the ring that only a single instance per zone is returned during quorum checks. The result of this is that there are `Ring.Get()` calls that _should_ fail when not enough zones are available that would now succeed since we return multiple instances per zone and the existing code only checks the number of instances, not zones.