@56quarters 2024-12-31
I'm going to close this PR. The changes to the ring required for https://github.com/grafana/mimir/issues/9944 are more complicated than expected. I'll put the various concerns into a design doc for some discussion with the Mimir team and other database teams as well.